requirements.txt
appdirs==1.4.3
Django>=3.0,<4.0
psycopg2-binary>=2.8
djangorestframework==3.12.4
Pygments==2.7.4
Markdown==3.3.4
django-filter==2.4.0
coreapi==2.3.3
dj-database-url==0.5.0
gunicorn==20.0.4
whitenoise==5.0.1
PyYAML==5.4
celery==5.1.2
django-celery-beat==2.2.1
feedparser==6.0.8
sqlalchemy==1.4.22
django-celery-results==2.2
Faker==8.11.0
factory_boy==3.2.0